+++ exinst/CHANGELOG.md
@@ -4,7 +4,15 @@
 
 * Re-export `Constraint` from `base`.
 
-* Add `Dict0` instance for `Bool`.
+* Add `Dict{0,2,3,4}` instances for `Bool`.
+
+* Add `GHC.Generics.Generic` support for `Some{1,2,3,4}`. This only works for
+  indexes with `PEnum` and `PBounded` instances.
+
+* Added tests.
+
+* Added `Test.QuickCheck.Arbitrary` instances for `Some{1,2,3,4}` in
+  `Exinst.Instances.QuickCheck`.
 
 
 # Version 0.2